[feat] minimap
미니맵 개발의 기본을 만들었다
