+++ .vscode/settings.json
@@ -56,5 +56,5 @@
         "temp/": true,
         "Temp/": true
     },
-    "dotnet.defaultSolution": "Unity_Project.sln"
+    "dotnet.defaultSolution": "Unity_Project_clone.sln"
 }
