--- Assets/RoomController.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 8b69e90ade1e1f84794e6116c637aa79
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
