+++ Assets/Scripts/Room.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 66f770b11d13bd24ea41beb4454aa7c1
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
