+++ Assets/Scripts/minimap.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: c0c31c1092e177b4a96d4ba2d3bc4bcc
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
