+++ Assets/Student/CSJ/prefab/minimaptile.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 135191ff71d545a498d97d11d5ca4713
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
